+++ test/functional-tests/javascript-obfuscator/JavaScriptObfuscator.spec.ts
@@ -1140,14 +1140,14 @@ describe('JavaScriptObfuscator', () => {
 
             const samplesCount: number = 30;
 
-            let areCollisionsExists: boolean = false;
-            let obfuscateFunc: (identifierNamesGenerator: TTypeFromEnum<typeof IdentifierNamesGenerator>) => string;
+            let collisionError: string | null = null;
+            let obfuscateFunc: (identifierNamesGenerator: TTypeFromEnum<typeof IdentifierNamesGenerator>) => IObfuscationResult;
 
             before(() => {
                 const code: string = readFileAsString(__dirname + '/fixtures/custom-nodes-identifier-names-collision.js');
 
                 obfuscateFunc = (identifierNamesGenerator: TTypeFromEnum<typeof IdentifierNamesGenerator>) => {
-                    const obfuscatedCode = JavaScriptObfuscator.obfuscate(
+                    return JavaScriptObfuscator.obfuscate(
                         code,
                         {
                             identifierNamesGenerator,
@@ -1156,9 +1156,7 @@ describe('JavaScriptObfuscator', () => {
                             identifiersDictionary: ['foo', 'bar', 'baz', 'bark', 'hawk', 'foozmos', 'cow', 'chikago'],
                             stringArray: true
                         }
-                    ).getObfuscatedCode();
-
-                    return obfuscatedCode;
+                    );
                 };
 
 
@@ -1167,10 +1165,14 @@ describe('JavaScriptObfuscator', () => {
                     IdentifierNamesGenerator.MangledIdentifierNamesGenerator
                 ].forEach((identifierNamesGenerator: TTypeFromEnum<typeof IdentifierNamesGenerator>) => {
                     for (let i = 0; i < samplesCount; i++) {
+                        const obfuscationResult = obfuscateFunc(identifierNamesGenerator);
+                        const obfuscatedCode = obfuscationResult.getObfuscatedCode();
+                        const seed = obfuscationResult.getOptions().seed;
+
                         try {
-                            eval(obfuscateFunc(identifierNamesGenerator));
-                        } catch {
-                            areCollisionsExists = true;
+                            eval(obfuscatedCode);
+                        } catch (error) {
+                            collisionError = `Seed: ${seed}. Error: ${error.message}`;
                             break;
                         }
                     }
@@ -1178,7 +1180,7 @@ describe('JavaScriptObfuscator', () => {
             });
 
             it('It does not create identifier names collision', () => {
-                assert.equal(areCollisionsExists, false);
+                assert.isNull(collisionError);
             });
         });